+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @talend/bootstrap-theme
 
+## 6.38.5
+
+### Patch Changes
+
+- 667cd0a50: chore: upgrade dependencies and align @talend scoped packages to latest
+- Updated dependencies [667cd0a50]
+  - @talend/icons@6.36.3
+
 ## 6.38.4
 
 ### Patch Changes